TN4 分支，一同走向新的一年

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/TippedDart.java
@@ -32,8 +32,8 @@ import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.wands.WandOfRegrowth;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.Crossbow;
+import com.shatteredpixel.shatteredpixeldungeon.journal.Catalog;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
-import com.shatteredpixel.shatteredpixeldungeon.plants.AikeLaier;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Blindweed;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Earthroot;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Fadeleaf;
@@ -42,7 +42,6 @@ import com.shatteredpixel.shatteredpixeldungeon.plants.Icecap;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Mageroyal;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Plant;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Rotberry;
-import com.shatteredpixel.shatteredpixeldungeon.plants.SkyBlueFireBloom;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Sorrowmoss;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Starflower;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Stormvine;
@@ -54,7 +53,7 @@ import com.shatteredpixel.shatteredpixeldungeon.windows.WndOptions;
 import com.watabou.utils.Reflection;
 
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.LinkedHashMap;
 
 public abstract class TippedDart extends Dart {
 	
@@ -78,13 +77,25 @@ public abstract class TippedDart extends Dart {
 	public void execute(final Hero hero, String action) {
 		super.execute(hero, action);
 		if (action.equals( AC_CLEAN )){
+
+			String[] options;
+			if (quantity() > 1){
+				options = new String[]{
+					Messages.get(this, "clean_all"),
+					Messages.get(this, "clean_one"),
+					Messages.get(this, "cancel")
+				};
+			} else {
+				options = new String[]{
+					Messages.get(this, "clean_one"),
+					Messages.get(this, "cancel")
+				};
+			}
 			
 			GameScene.show(new WndOptions(new ItemSprite(this),
 					Messages.titleCase(name()),
 					Messages.get(this, "clean_desc"),
-					Messages.get(this, "clean_all"),
-					Messages.get(this, "clean_one"),
-					Messages.get(this, "cancel")){
+					options){
 				@Override
 				protected void onSelect(int index) {
 					if (index == 0){
@@ -94,7 +105,7 @@ public abstract class TippedDart extends Dart {
 						hero.spend( 1f );
 						hero.busy();
 						hero.sprite.operate(hero.pos);
-					} else if (index == 1){
+					} else if (index == 1 && quantity() > 1){
 						detach(hero.belongings.backpack);
 						if (!new Dart().collect()) Dungeon.level.drop(new Dart(), hero.pos).sprite.drop();
 
@@ -122,6 +133,7 @@ public abstract class TippedDart extends Dart {
 		if (durability <= 0){
 			//attempt to stick the dart to the enemy, just drop it if we can't.
 			Dart d = new Dart();
+			Catalog.countUse(getClass());
 			if (sticky && enemy != null && enemy.isAlive() && enemy.alignment != Char.Alignment.ALLY){
 				PinCushion p = Buff.affect(enemy, PinCushion.class);
 				if (p.target == enemy){
@@ -154,37 +166,39 @@ public abstract class TippedDart extends Dart {
 	@Override
 	public float durabilityPerUse() {
 		float use = super.durabilityPerUse(false);
-		
-		use /= (1 + Dungeon.hero.pointsInTalent(Talent.DURABLE_TIPS));
 
-		//checks both destination and source position
-		float lotusPreserve = 0f;
-		if (targetPos != -1){
-			for (Char ch : Actor.chars()){
-				if (ch instanceof WandOfRegrowth.Lotus){
+		if (Dungeon.hero != null) {
+			use /= (1 + Dungeon.hero.pointsInTalent(Talent.DURABLE_TIPS));
+
+			//checks both destination and source position
+			float lotusPreserve = 0f;
+			if (targetPos != -1) {
+				for (Char ch : Actor.chars()) {
+					if (ch instanceof WandOfRegrowth.Lotus) {
+						WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
+						if (l.inRange(targetPos)) {
+							lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
+						}
+					}
+				}
+				targetPos = -1;
+			}
+			int p = curUser == null ? Dungeon.hero.pos : curUser.pos;
+			for (Char ch : Actor.chars()) {
+				if (ch instanceof WandOfRegrowth.Lotus) {
 					WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
-					if (l.inRange(targetPos)){
+					if (l.inRange(p)) {
 						lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
 					}
 				}
 			}
-			targetPos = -1;
+			use *= (1f - lotusPreserve);
 		}
-		int p = curUser == null ? Dungeon.hero.pos : curUser.pos;
-		for (Char ch : Actor.chars()){
-			if (ch instanceof WandOfRegrowth.Lotus){
-				WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
-				if (l.inRange(p)){
-					lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
-				}
-			}
-		}
-		use *= (1f - lotusPreserve);
 
 		float usages = Math.round(MAX_DURABILITY/use);
 
 		//grants 3+lvl extra uses with charged shot
-		if (bow != null && Dungeon.hero.buff(Crossbow.ChargedShot.class) != null){
+		if (bow != null && Dungeon.hero != null && Dungeon.hero.buff(Crossbow.ChargedShot.class) != null){
 			usages += 3 + bow.buffedLvl();
 		}
 
@@ -201,22 +215,20 @@ public abstract class TippedDart extends Dart {
 		return 8 * quantity;
 	}
 	
-	public static HashMap<Class<?extends Plant.Seed>, Class<?extends TippedDart>> types = new HashMap<>();
+	public static final LinkedHashMap<Class<?extends Plant.Seed>, Class<?extends TippedDart>> types = new LinkedHashMap<>();
 	static {
-		types.put(Blindweed.Seed.class,     BlindingDart.class);
-		types.put(Mageroyal.Seed.class,     CleansingDart.class);
-		types.put(Earthroot.Seed.class,     ParalyticDart.class);
-		types.put(Fadeleaf.Seed.class,      DisplacingDart.class);
-		types.put(Firebloom.Seed.class,     IncendiaryDart.class);
-		types.put(Icecap.Seed.class,        ChillingDart.class);
 		types.put(Rotberry.Seed.class,      RotDart.class);
-		types.put(Sorrowmoss.Seed.class,    PoisonDart.class);
-		types.put(Starflower.Seed.class,    HolyDart.class);
-		types.put(Stormvine.Seed.class,     ShockingDart.class);
 		types.put(Sungrass.Seed.class,      HealingDart.class);
+		types.put(Fadeleaf.Seed.class,      DisplacingDart.class);
+		types.put(Icecap.Seed.class,        ChillingDart.class);
+		types.put(Firebloom.Seed.class,     IncendiaryDart.class);
+		types.put(Sorrowmoss.Seed.class,    PoisonDart.class);
 		types.put(Swiftthistle.Seed.class,  AdrenalineDart.class);
-		types.put(SkyBlueFireBloom.Seed.class,  HaloDart.class);
-		types.put(AikeLaier.Seed.class,  LightDart.class);
+		types.put(Blindweed.Seed.class,     BlindingDart.class);
+		types.put(Stormvine.Seed.class,     ShockingDart.class);
+		types.put(Earthroot.Seed.class,     ParalyticDart.class);
+		types.put(Mageroyal.Seed.class,     CleansingDart.class);
+		types.put(Starflower.Seed.class,    HolyDart.class);
 	}
 	
 	public static TippedDart getTipped( Plant.Seed s, int quantity ){
